Rename Map to EntryMap, move former FilteredList functionality to EntryList, extract NotifyMe component, improve translation usage
